Add cluster-pool-ipv4-min-candidate-index to cilium operator
By default the candidate index starts at 0. This flag allows us
to set a different value for the min. That implies that Node CIDRs
are allocated from a higher range.

Ex:
Without this flag, the node ranges look like this
10.0.0.0/24 - node1
10.0.1.0/24 - node2
...

With this flag (value 256), we can configure node ranges to
start at a higher value
10.1.0.0/24 - node1
10.1.1.0/24 - node2
10.1.2.0/24 - node3
...